==> step6.2_prediction_knn.py <==
# IMPORT BOUNDARIES
import pandas as pd
import sys
from sklearn import metrics
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split
from sklearn.neighbors import KNeighborsClassifier
from sklearn.metrics import confusion_matrix
from sklearn.model_selection import GridSearchCV

# READ DATA
df = pd.read_csv("datasets/preprocessing5_balance_gender/balanced_gender.csv", error_bad_lines=False, header=0, sep=';', low_memory=False)

# SPLIT DATA IN TRAIN AND TEST SET
y = df['ANREDE']
X = df.drop(['ANREDE'], axis=1)
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=1)

# SCALING DATA
sc = StandardScaler()
X_train = sc.fit_transform(X_train)
X_test = sc.transform(X_test)

# NEXT PART SHOULD NOT BE DONE EVERY TIME
'''''''''
# SEARCHING FOR BEST PARAMETERS
grid_params = {'n_neighbors': [3,5,7,9,11],
               'weights': ['uniform', 'distance'],
               'metric': ['euclidean', 'manhattan']}

gs = GridSearchCV(KNeighborsClassifier(),
                  grid_params,
                  verbose=1,
                  cv=3,
                  n_jobs=-1)

gs_results = gs.fit(X_train, y_train)

print(gs_results.best_score_)
print(gs_results.best_estimator_)
print(gs_results.best_params_) #n_neighbors=9, weights='distance', metric='manhattan'
'''''''''

# CREATE KNN CLASSIFIER WITH BEST PARAMETERS FOUND BY GRIDSEARCH
knn = KNeighborsClassifier(n_neighbors=9, weights='distance', metric='manhattan')

# TRAIN AND PREDICT
knn.fit(X_train, y_train)
predictions = knn.predict(X_test)

# PRINT MODEL ACCURACY
print("Erfolgsrate des kNN Algorithmus: ", round(metrics.accuracy_score(y_test, predictions), 3) * 100, "%")

# ECHO OUTPUT IN FILE
sys.stdout = open("c:\\Projekt\projekt_geschlecht2\Echo_files\knn.txt", "w")
print("k-NearestNeighbor got an accuracy score of ", round(metrics.accuracy_score(y_test, predictions), 3) * 100, "%.")
